Enforced use of type OpCode over uint8_t.
The OpCode type was commonly used as uint8_t in some areas. The enum
OpCode was restricted into the size of uint8_t and areas with pointers
being used were reinterpret_cast into uint8_t if arithmetic was needed.

The benefit is being able to see the defined OpCode value during debugging instead of
an unsigned integer value.