Added Zura Exit codes and ZuraExit function.
Exit codes can be helpful to know why a program ended. When the
interpreter finds some kind of fault and calls exit(), the resulting
exit codes haven't been very helpful. Now we can use a list of values to
as a reference for to help understand why the interpreter faulted.