New Makefile adds support for GUI.
A new Makefile can move forward in supporting a GUI using the DearImgui
library. The library will use GLFW for events and basic window
management. OpenGL3 is the API used for drawing the GUI.

This opens the possiblity of making tools native to the development of
Zura and using Zura. For example, a console dialog for the REPL that can also display
values in the VM or displaying mathematical function graphs.

DearImgui can  with the source code. GLFW will need to be resolved
as a dependency for the developer. See https://www.glfw.org/download.