SRC_DIRS variable used.
This variable keeps a running list of the directory structure of the
project. If an additional directory is created in the ./src/ directory,
it must be added to the list. This allows the project to create .obj
files into a dedicated ./obj/ directory and keep them from littering
the ./src/ directory tree when developing.